1.opt sender.

+++ socket/protocol/tcp.go
@@ -42,7 +42,7 @@ func (d *DefaultTcpProtocol) IsUnknown(messageType byte) bool {
 	return messageType == Unknown
 }
 
-func (d *DefaultTcpProtocol) Decode(message []byte) (messageType byte, code int, id int64, route []byte, body []byte) {
+func (d *DefaultTcpProtocol) Decode(message []byte) (messageType byte, code uint32, id uint64, route []byte, body []byte) {
 	if !d.isHeaderInvalid(message) {
 		return 0, 0, 0, nil, nil
 	}
@@ -54,12 +54,12 @@ func (d *DefaultTcpProtocol) Decode(message []byte) (messageType byte, code int,
 	headLen := d.HeadLen()
 
 	return message[2],
-		int(binary.BigEndian.Uint16(message[8:12])),
-		int64(binary.BigEndian.Uint64(message[12:headLen])),
+		binary.BigEndian.Uint32(message[8:12]),
+		binary.BigEndian.Uint64(message[12:headLen]),
 		message[headLen : headLen+int(message[3])], message[headLen+int(message[3]):]
 }
 
-func (d *DefaultTcpProtocol) Encode(messageType byte, code int, id int64, route []byte, body []byte) []byte {
+func (d *DefaultTcpProtocol) Encode(messageType byte, code uint32, id uint64, route []byte, body []byte) []byte {
 	switch messageType {
 	case Bin:
 		return d.packBin(code, id, route, body)
@@ -164,7 +164,7 @@ func (d *DefaultTcpProtocol) getLen(message []byte) int {
 	return rl + int(bl) + headLen
 }
 
-func (d *DefaultTcpProtocol) packBin(code int, id int64, route []byte, body []byte) []byte {
+func (d *DefaultTcpProtocol) packBin(code uint32, id uint64, route []byte, body []byte) []byte {
 
 	var rl = len(route)
 
@@ -191,10 +191,10 @@ func (d *DefaultTcpProtocol) packBin(code int, id int64, route []byte, body []by
 	binary.BigEndian.PutUint32(data[4:8], uint32(bl))
 
 	// 8 - 11 code
-	binary.BigEndian.PutUint32(data[8:12], uint32(code))
+	binary.BigEndian.PutUint32(data[8:12], code)
 
 	// 12 - 19 id
-	binary.BigEndian.PutUint64(data[12:headLen], uint64(id))
+	binary.BigEndian.PutUint64(data[12:headLen], id)
 
 	copy(data[headLen:headLen+rl], route)
