+++ socket/protocol/udp.go
@@ -77,7 +77,7 @@ func (d *DefaultUdpProtocol) HeadLen() int {
 	return 20
 }
 
-func (d *DefaultUdpProtocol) Decode(message []byte) (messageType byte, code int, id int64, route []byte, body []byte) {
+func (d *DefaultUdpProtocol) Decode(message []byte) (messageType byte, code uint32, id uint64, route []byte, body []byte) {
 	if !d.isHeaderInvalid(message) {
 		return 0, 0, 0, nil, nil
 	}
@@ -89,12 +89,12 @@ func (d *DefaultUdpProtocol) Decode(message []byte) (messageType byte, code int,
 	headLen := d.HeadLen()
 
 	return message[2],
-		int(binary.BigEndian.Uint16(message[8:12])),
-		int64(binary.BigEndian.Uint64(message[12:headLen])),
+		binary.BigEndian.Uint32(message[8:12]),
+		binary.BigEndian.Uint64(message[12:headLen]),
 		message[headLen : headLen+int(message[3])], message[headLen+int(message[3]):]
 }
 
-func (d *DefaultUdpProtocol) Encode(messageType byte, code int, id int64, route []byte, body []byte) []byte {
+func (d *DefaultUdpProtocol) Encode(messageType byte, code uint32, id uint64, route []byte, body []byte) []byte {
 	switch messageType {
 	case Bin:
 		return d.packBin(code, id, route, body)
@@ -162,7 +162,7 @@ func (d *DefaultUdpProtocol) getLen(message []byte) int {
 	return rl + int(bl) + headLen
 }
 
-func (d *DefaultUdpProtocol) packBin(code int, id int64, route []byte, body []byte) []byte {
+func (d *DefaultUdpProtocol) packBin(code uint32, id uint64, route []byte, body []byte) []byte {
 
 	var rl = len(route)
